feat: optionally enforce muteOnStart on dial-in participants
The `muteOnStart` state is currently enforced on all participant types
(web or dial-in/SIP). Depending on the usage, it might no be desirable
to enforce it on dial-in endpoints due to:
  - Voice being their only mode of communication
  - The ability to unmute via DTMF tones not being clear enough to the
    end user

Introduce an akka-apps  config `voiceConf.dialInEnforceMuteOnStart` that
controls whether the meeting mute state is applied to dial-in at join time.

Default behavior is now changed to *not enforcing mute state for
dial-in* on join (i.e.: false, always join unmuted). The rationale is that
dial-in users joining muted was never an explicit goal of anyone. It's just
an unplanned side-effect of setting muteOnStart=true by default.