fix(livekit): mitigate audio publish timeouts from overlapping calls
LiveKit has an issue where overlapping publish calls for the same
track type cause an inconsistent server state. This leads to all
subsequent publish calls (of the target track type) timing out due
to a race condition in the offer/answer process with cancelled calls.

End users experience automatic audio muting when the timeout occurs.
The only way to remediate this inconsistent state is by re-creating
the participant's LK connection.
This issue is most often hit in meetings with muteOnStart=false or
when participants reconnect as unmuted.

Introduce a publish/unpublish processing queue in the LK audio
bridge to prevent overlapping publish calls. Serialize requests,
and skip overlapping ones whenever possible. This greatly reduces
the occurrence of the timeouts, based on field trial deployments of this
patch.